driver: add framebuffer backed console
We're in graphics mode and we are drawing the font.
Upside of this is we can have high resolution (like 4K) and nicer fonts,
downside is obvious too that it seems really slow on real hardware,
although qemu seems OK with it (fb simulated with memory). On my 4K
monitor, the refresh rate is about 2 fps.

We have used the double buffer. Future optimizations:
1) Enable write-combining memory for framebuffer (10x faster)
2) rep movsb
3) Maybe a graphics driver?